fix(syncofficers, officers): add helper functions for officers
